+++ .github/CHANGELOG.md
@@ -17,7 +17,7 @@ This is the changelog for the version V2.6.0. Don't forget to [join us on Discor
 * Don't modify ammo on Kit-Pickup
 * simplify permissions. No need to be on the server anymore
 * REAL AIMING OF BOTS Thanks to Bree_Arnold!!! Please use the dev-version of VU
-* Recoil-Compensation added for all Bots
+* Recoil-Compensation added for all Bots dependant of skill
 * keep Skill of Bot for whole game
 * restructurisation of Bot.lua-Class